FileConfigurationTest to TextConfigurationTest
